+++ yash-syntax/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `impl FromStr for parser::lex::Keyword`
 - `parser::lex::Operator::as_str`
 - `parser::lex::ParseOperatorError`
+- `parser::lex::TryFromOperatorError`
 
 ### Changed
 
@@ -21,7 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `parser::lex::Operator::is_clause_delimiter` now `const` and `#[must_use]`
 - `<parser::lex::Operator as FromStr>::Err` from `()` to `ParseOperatorError`
 - `<syntax::AndOr as FromStr>::Err` from `()` to `ParseOperatorError`
+- `<syntax::AndOr as TryFrom<parser::lex::Operator>>::Error` from `()` to `TryFromOperatorError`
 - `<syntax::RedirOp as FromStr>::Err` from `()` to `ParseOperatorError`
+- `<syntax::RedirOp as TryFrom<parser::lex::Operator>>::Error` from `()` to `TryFromOperatorError`
 
 ### Removed